backend/requirements.txt
fastapi==0.118.3
uvicorn==0.37.0
aiofiles==25.1.0
python-multipart==0.0.20
openai==2.3.0
python-dotenv==1.1.1
pillow==11.3.0
pymongo[srv]==4.15.4
cloudinary==1.41.0
